pyproject.toml
[project]
name = "datagovmy-ai"
version = "1.0"
dependencies = [
    "fastapi",
    "langchain",
    "lanarky",
    "faiss-cpu",
    "weaviate-client",
    "dagster",
    "dagster-webserver",
    "pandas",
    "openai",
    "pyarrow",
    "tiktoken",
    "redis",
    "fastapi-cache2[redis]",
    "uvicorn",
]

[project.optional-dependencies]
dev = ["pytest", "jupyter", "locust", "pre-commit"]
